feat: Remove "& Sangha"
Personal choice: we can discuss at the meeting tomorrow.

The original intention behind "& Sangha" was that it would involve text
written by various sangha members, but that was back when it was called
the "Handbook" and had a different intention. As it stands currently,
all text is written by Dawai, so it makes sense for him to be the
author. We can credit the sangha on a separate page for help with
research, quotes, feedback, etc.